Refactoring of preferences pages generation
OB03-036

Code refactoring for preferences pages generation using the new
Preferences_Page_Interface API defined in default_preferences.ads.

This refactoring is done to facilitate the comprehension of the
pages generation code and to facilitate the integration of
widgets created out-of-the box by other modules
(e.g: Key Shortcuts, Color Theme) in the GPS preferences dialog.

Change-Id: Icedcf7e0c32cf3d3f2e0eced4ff654bfb346e6be