The GPS window title is now configured via a preference
We no longer hard-code its contents (GPS - base name - dir name - project),
but instead use a new preference and macro expansion to let users
configure it. One of the minor changes is that we might end up with extra
" - " in the title when parts of it do not apply to the current context.
We also no longer display the name of the remote host.

This change also applies to floating windows, for which users could not
control the title before.

"%fd", "%ts", "%tl", "%rbl": new macros

N714-016 (1/2)

Change-Id: I36fc6a9d5d67fcb54a5822f0c56153d86bc62cde