Saving a new file might impact the internal name of existing editors
When a file is opened in an editor (prefix/src/a.adb for instance), and
we create a new file in (prefix), then save it, this would impact the
internal for existing editors, because we are passing a directory
(prefix) as the original to the File_Renamed_Hook. As such, the existing
buffer is then saved in prefix/adb from then on.

The solution is not to emit File_Renamed_Hook when saving a new file.

At the same time, a cleanup of the code for the hook was done, including
adding more comments to make the code clearer.

OA15-027

Change-Id: I6815e3496db85e256e921c50256c9874e868d24d