Do not create actions for the menus anymore
We used to create one action for each menu that is created. This costs
startup time (we are creating several hundreds menus with all the /Help
menus), memory to store these actions, and complexifies the code.

As a result, we could also remove Register_Perma_Command, a left over
from the old days, as well as several other functions. The keymanager
is simplified since it doesn't need to manipulate menus directly.

Backward compatibility is assured: GPS.execute_action("/menu") still
works (and wasn't using those actions, in fact it was creating a new
command and never freeing it, a minor memory leak); entries in keys.xml
that reference menus are automatically associate with the corresponding
action instead now.
This might not work for menus that are not associated with actions, though,
like the /Window/ menus (there are corresponding actions though, and this
menu will be changed as part of the work on system menus in N829-014).

N829-014

Change-Id: I9ab0e6cdcbff925cda48e2d78fd2c7d6d191de03