GPS.Menu: no longer derives from GPS.GUI
A menu is no longer necessarily a graphical element, since it could be
displayed outside of the application. This is better too, since rather
than disabling or executing a menu, it is better to do it on the
associated action directly. A compatibility layer is provided in
extensions/__init__.py, except for pywidget() and destroy()

A GPS.Action also no longer derives from GPS.GUI, and it should never
have in fact since it never made sense.

(Register_Menu): remove the subprograms that manipulate Gtk_Menu_Item
from the spec, they are no longer used by the rest of GPS. Minor
performance update when creating the initial menubar, by avoiding extra
tests and path manipulation that are not necessary in this context.

pygps.WidgetTree now returns the widgets in the same order as gtk+
(i.e. the visual order), rather than the reverse order. This impacts a
number of tests, but was necessary to get the menus in the right order.

OB13-048  (1/2)

Change-Id: I600c91ca8048ed8c4f585ac6127244fa152d01f7