Plugins search provider integrated in preferences/omni search
OB23-033

A newly created search provider for plugins has been integrated
in the omnisearch and the local preferences search bar.

The GPS.Search and GPS.Search.GUI packages have been refactored to
be able to use and extend Overall_Search_Provider type, used in the
omnisearch, in other packages (e.g: in GPS.Kernel.Preferences_Views
to create the preferences dialog local search bar).

The Startup_Module package files have been moved in the 'kernel'
directory, allowing us to remove the 'theme_manager' subproject.

Some refactoring has been made in the packages related with the
preferences search and the preferences dialog so that preferences
highlighting works with the new way of displaying the plugins
(in one page instead of one for each plugin).

Change-Id: Idd34bc3577c26b9245724636b60ec204a9e575d1