Add a menubar to floating editors and browsers
Fix initial title of floating windows (which was the MDI's default, until
a child_selected signal had been emitted, which could be much later if
the child was already the focus child).

Floating windows for editors and browsers are now created as
GPS_Application_Window (a new type), so that they have a special handling
of their menubar and those are properly refreshed when the context changes.

Fix minor memory leak is Is_Any_Menu_Open. This function now looks into
all known menu bars, not just that of the main window.

(gps-main_window-debug): remove special handling of the Protection Domains,
which is now handled via the usual action filters. This makes it compatible
with contexual menus, system menus, ...

G222-043

Change-Id: Ie910613931b9fe735ebfec4543cfe414394e9bbe