Create menus from Gmenu_Model
We no longer create the menus directly from the menus.xml. Instead,
we always go through a Gmenu_Model, which is necessary for system
menus (N829-014), but now also used when we create the menu bar
ourselves. This is slightly faster (since we avoid the handling of
Add_Before, and always append to the menus), and more importantly
allows us to create the meny model in memory.
That way, any call to Register_Menu modifies the menu model, so that
when we create toolbars for floating windows (G222-043), they contain
the menus added from python or the dynamic menus.
No longer reparse menus.xml when opening a floating editor, since the
menu model is already stored in memory.

(Register_Menu): add the new item to all menubars, including floating
windows.

(Create_Menu_Path): fix handling of mnemonics. We were removing all
undescore characters, even when we had "__" (which should be a single
underscore). As a result, some menus like /Build/*/* were not
correctly associated with a menu path when the makefile target had
underscores.

Change how the builder module cleans up its menus: we now unregister
action (as we did before), which automatically removes associated
menus (optionally). This removes direct manipulations of widgets
in the builder model, and calls to Find_Menu_Item (which no longer
works now that there can be multiple menu items for a given path)

Change-Id: I3ea2cbaf453cf0c543da913b6dafd0e70cfed875