Fix the order of preferences in the dialog
OB30-039

Preferences are now displayed in the same order as they were created.

The disorder was caused by the fact that we stored the preferences
in Indefinite_Hashed_Maps, which don't preserve the order.

Using Dounly_Linked_Lists instead of Indefinite_Hashed_Maps
preserves the order when appending new preferences to a
given page.

Change-Id: I01df724956f3ad91386e3f35824d34361aca8448