(Launch_Foreground_Command): removed
(Execute_In_Background): renamed to Execute_Action.

Remove all calls to this subprogram. Instead, additional capabilities
were given to Execute_Action, with the goal ultimately that this should
be the only way to execute an action. We had two places where we were
creating proxys and then launching the action's command, and without
having an explicit trace in the log file, so it was sometimes
difficult to know which actions are executed by the user.
This should help debug things via the log file.

Found while working on G222-043

Change-Id: Ic087c8aef1b296b9a4a572eadd5058e5e4c525f3