Cleanup handling of initial focus when opening editors
We were running the context_changed hook three times, and often
with invalid information. We were trying to give the focus to the
editor before it had been inserted in the MDI. This resulted in
a context_changed running with the old child as the context.

We were then running context_changed once because we were raising
the window, and then one last time after setting the initial
location.

This also seems to partially fix the filepos.py plugin: the previous
position in the editor is correctly restored, although the editor
is not scrolled to show it.

Minor fixes in vcs module to avoid code duplication (found while
looking at unrelated regressions in previous versions of this commit)

In the context of G222-043

Change-Id: I6b4492e49fa10e0d17fd417eecd09e2f15a841cd