(Register_Menu): update the menu model
The effect is that system menus now show dynamic menus inserted after
menus.xml was parsed, and when not using system menus, floating windows
also have those menus in their own menubar at startup (it is already
the case that existing menubars are updated when a new menu is
registered)

Store action->menu associations in a separate data structure, so that it
can be filled even before actions are created. This is used for system
menus in particular, for which we have no clear point when to call
Lookup_Action (which was so far in charge of updating the list of menus
for all the actions).

Remove use of Find_Menu_Item for /Debug/Initialize, /File/Recent and
/Project/Recent.
This ensures that these menus are also visible in floating windows or in
system menus (N829-014)

G222-043 (1/2)

Change-Id: I9ec84e9ec743ec58a870488ce699b3808b49bd18