Put back the /Build/Recompute Xref Info menu
This menu was associated with the "recompute xref info" action, which is
generated on the fly from the homonym build target. Then the builder
module would unregister this action (which now removes the associated
menus) whenever we parse the targets again. As a result, the menu no
longer showed up after the previous patch.

So we now use an intermediate action associated with the menu.

G222-043

Change-Id: Ifed705b0d14a53ff3ee6c1aa28229f127a8e293c