fix various valgrind issues with commands
A Monitor_Command (for an external process) is wrapped inside a
Scheduled_Command (for the task manager). When the scheduled task
is freed, the monitor command is also freed, but it has a reference
to the scheduled command (used to print a message on the console).
This results in access to freed memory. Remove this internal pointer,
and now pass the low-level command explicitly to the various output
filters (they can still access the scheduled command via a new
Get_Scheduled_Command function in the kernel)

(Next): when moving to the next command in a command list, we should
not free the first command before we have removed it from the list.
Otherwise, the task manager will find that the command we are freeing
belonged to the queue and will try to execute the next command (thus
a nested call to Next, which also results in access to dangling
memory).

(Free): renamed to Primitive_Free, for commands, to make sure the code
is not calling this directly. This fixes various cases where Ref/Unref
was not working properly for commands.

(Execute_Incremental): fix case where we could report a Queue_Changed
event on the wrong queue.
(Active_Incremental, Passive_Incremental): in case of exception, we could
be left in a state where the task manager is not running, but GPS believes
it is.
(GPS.Kernel.Timeout): remove the timeout loop in this package, since the
task manager already has one. This is more efficient, and clarifies the
code (avoids reentrency in some cases). Make sure Exit_Cb is only run
once.

OC14-005

Change-Id: I9972f937d21388e4ba5115c2773d09950a4c567a