Prevent sawing-the-branch-you're-on in the task manager
It could happen that a command could tell the Task Manager to
Interrupt its own queue - this in turn could result in the command
to be freed, and the data it is trying to access be invalid.
Prevent this by only Freeing queues at the start of a task manager
run.

For OC14-005

Change-Id: I80d538917aed34643cc3772b599861bdddbd206f