(shell.py): automatically emit "stty echo"
This is more user-friendly, and should work in most cases. When it
doesn't, there is now a preference to disable this behavior. We also
provide an environment variables that users can test in their .bashrc to
set specific prompt or others, for GPS.

Add support for addition ANSI escape sequences needed by vi and the
shell (DELETE_CHARS, in particular, but also basic and incomplete
support for scrolling regions, or INSERT_ONE_LINE).

OC17-020

Change-Id: I047d99d0024721ec519958d67a1391dc329508fc