M412-032 Enable analysis of generics from GPS menus
SPARK menus in GPS now correctly set -U when attempting analysis of a
subprogram inside a generic unit (both flow analysis and proof) or proof
of a line inside a generic unit. Ideally we'd like this to also be enabled
on generics inside non-generic units, but this is not easy to do with
current information in GPS, so delay that refinement for a future time
when this might be easier.

(on_prove_line, on_prove_check): Add -U when inside generic unit.
(generic_action_on_subp): Add -U when inside generic unit.
(mk_debug_loc_string): New function to retrieve location for debugging
  purpose.
(inside_generic_unit_context): New function to determine when inside a
  generic unit.

Change-Id: I5e265adf818a70825ccc472791fbf47665b02b10