tip_of_the_day.py moved to /share/support/core
OC21-021

The plugin file providing the "Tip of the Day" feature has
been moved to /share/support/core since that this feature
can be enabled/disabled from the "General" preferences page.

Thus, it must not appear in the "Plugins" preferences page list
too: otherwise, there will be two different ways to enable/disable
this feature.

The 'Tip of the day number' preference is now set as hidden, since
that its value can already been modified by clicking on the arrow
buttons of the tip window.

Change-Id: I3e87426e0f89359dc7b190a197b679baf8ca78ff