"Editor/Fonts & Colors" preferences page reordered
OC21-021

The "Editor/Font & Colors" preferences page has been reordered.

The Python API used to register new styles for highlighters
has been slightly modified so that we can associate a label and
more specific documentation for this new style, in order to have
more readable preferences.

Moreover, the style preferences associated with style created with
this API are now registered in the "Editor/Fonts & Colors" page
instead of creating a specific page for each language.
This is done in order to have only one page for all the
fonts/colors related preferences.

Change-Id: I8e7a1070bd34605ac75f48d8d1fb1f9085b3a63a