Use of flow boxes for preferences groups
P113-022

The preferences inside groups are now displayed using Gtk_Flow_Box
widgets.

This has the effect to reduce the size of all the preferences pages
allowing the user to see directly almost all the displayed preferences
for a given page when the preferences dialog is displayed with the
default size.

Checboxes are now right aligned to, with the check button on the
right side of the corresponding label.

Change-Id: I73af8855383adc2375d610e1bb3783dd502c39b0