Advanced preferences page refactoring
P113-022

An entry in the preferences dialog tree view is now always
created for the advanced preferences page.

The user can decide whether he wants to show this entry or
not using the preferences dialog local menu

The dialog preferences search provider only searches among
hidden preferences when the advanced preferences page is
currently shown in the dialog.

Change-Id: I3b43cc49f04b0f8947ff263a49dff95238f87adc