Remove 'Preferences dialog' documentation in Sphinx
This was not up-to-date (missing lots of preferences, and still
documenting onces we have removed a long time ago).

Review the documentation for almost all preferences to make them
shorter, now that they are displayed in the Preferences dialog.

Change the type of the Default Font preference (which was a style,
not a font), since we never used foreground and background.

Reorganize some of the preferences into groups.

Remove a few preferences that were unused

P113-026

Change-Id: I31a7381b0637d199ac0ec503878a96affec437e6