P122-018 (1/2) Give focus to the preferences dialog search bar
When opening the preferences dialog, the focus is now
given to the search bar.

This has been done at the Generic_Views package level,
so that every generic view that contains a search/filter
bar and that does not ask to give focus to a specific
child widget (via the Initialize function) actually
gets its search/filter bar receiving it.

Change-Id: I33552e5a1fcac1c62848f93a1e066c4c473fa8e7