P108-047 Correction of use Containing_Directories
As described in AARM:

Discussion: This is purely a string manipulation function.
If Name is not given as a full name, the containing directory
probably won't be one, either. For example,
if Containing_Directory ("..\AARM\RM-A-8") is called on Windows®,
the result should be "..\AARM". If there is no path at all
on the name, the result should be "." (which represents the
current directory). Use Full_Name on the result of
Containing_Directory if the full name is needed.

Change-Id: I10742473c98c142f87f96aaf21933d19bcd29c41