Do not execute python actions after GPS has terminated python
This might happen when a process (like the constructs) was still
running when the user exits GPS.

Terminate the idle and timeout loops from the task manager when
GPS terminates, since otherwise they might happen to run after
we have destroyed parts of the kernel.

Fix a few errors reported by valgrind on exit

Add documentation and ??? comments to Task_Manager.Run.

P222-005

Change-Id: Ic7dba295015901fdd3846409ff40701931e96c4d