Reimplement the Task Manager HUD in Python
This is a step towards simplification of the implementation of the
task manager, by separating the GUI from the engine.

The GUI is now implemented entirely in Python, and relies only on the
GPS.Task API, rather than being implemented within the actual task
manager engine.

Make the HUD button pop up an ephemeral task manager window, which
allows users to view and interrupt tasks without having to bring
the complete task manager view into their MDI.

For P314-001.

Change-Id: Ib16ffbcbed58a56f5974c67283d080ac4a002cf5