Fix the 'Debug/Connect to board...' feature
P310-040

When initializing a debug session in non-native mode, the user
is now able to connect to a remote target if a gdb-server
has been spwaned for the protocol and the target retrieved
from the Project properties.

The Connect_To_Target procedure has been modified so that
it does not freeze the UI if no gdb-server has been spawned
for this protocol/target.

Change-Id: I629e6cffe7c1ab54288a62e22df2b9decaaf9e02