Further isolate GVD modules from each other
Inter-views dependencies should be limited as much as possible.  This
reduces risks of elaboration circularities, and makes it easier to
maintain the views or add new ones.

Rework the various subprograms that send commands to the debugger, to
share code and fix some issues. For instance, "graph *" commands were a
special case (done in multiple places) even though we now have a hook to
let modules and plug-ins handle user commands. But such user commands
were not processed correctly for commands that had been queued because
another one was already running when they were send initially.  As a
result of the cleanup, there is no longer a need for the special case of
the graph commands, and the GVD.Canvas module is independent.  Similar
processing will be needed for the variables view.

Rename the "Send" subprograms to avoid name clashes and make the code
easier to read.

Part of OC30-003

Change-Id: I5238ce5e29146f2f555b24fdbcb8dbc827b9dc51