Remove nested scrolled window in breakpoints editor
We had both a general scrolled window for the whole view, and another
one for the list of breakpoints, which makes it hard to view all
breakpoints with the mouse scroll.

Rewrite the code to build the GUI, which had been automatically
generated a long time ago, but can be made more readable to ease
maintenance.

P315-043

Change-Id: If0c4fc316e8e07b9c98ac3711e28c9e219441550