Revamp the breakpoints view
Use a local toolbar for Remove/View/Advanced, instead of adding large
buttons at the bottom of the view. For the toolbar, new GPS actions
were defined, so key bindings can be associated to them.

The view now only contains the list of breakpoints. Toolbar buttons open
a dialog that shows all properties of the breakpoint (basic properties
like file, line, exception,... as well as advanced properties like
conditions and commands).
The previous workflow of showing the properties of the selected breakpoint
and letting users edit them to create new breakpoints could be misleading
(for instance, selecting a breakpoint, then changing the line number would
create a new breakpoint, and not edit the selected one).

The view is thus smaller now and easier to keep permanently in the desktop.

P315-043 (1/2)

Change-Id: Ie478f915c7f3eb0cb2a7a6beb4805d88c4fcb0ef