New view: debugger variables
This shows the same information as the data window, but in a tree,
which a lot of users find more convenient

Improve display of <repeat> types, so that there are 2 less nodes
to expand to see the actual values.

Add support for splitting the debugger output (`cmd`) into several
components, rather than consider it as a single block of text.

Gtkada.Tree_View: add support for writing simple filters and showing
parents when children should be visible.

Local configuration: whether to show types or not.

Display in red modified values between two debugger actions.
Preserve expansion of nodes between refresh of the tree. This is done
 in Gtkada.Tree_View in a generic manner, and can be reused for other
 views.

Add support for editing values of variables from the Variables view.
 This required making the gdb "set variable" command run the hooks so
 that the view is properly refreshed.

Change default debug perpsective, to replace the Data window with the
 Variables view. It is now at the bottom, rather than take valuable
 space at the top.

Access types can be expanded in place to see the referenced value. This
 is done by new general support in Gtkada.Tree_View for lazy nodes,
 which can be used in other contexts.

Remove contextual menu Debug/Print A, which only emits a "print ..."
 in the console and is therefore not very useful. Rename
 /Debug/Display A to /Debug/Graph Display A, since we now also have
 /Debug/Tree Display A. Move the graph commands to gvd-canvas.adb for
 better separation of modules.

OC30-003 (1/2)

Change-Id: I567f0b23573cccef53f15e8db0c0b803e21235d2