Create a view to display memory usage
N701-033 (1/2)

Spawn a view that displays the memory usage for each memory
region of an executable.

This view is spawned only if the linker supports the
'--print-memory-usage' switch: this switch has been
added on the 'Build All' and 'Build Main' build targets
command lines and is disabled by a switch filter if the
currently used linker does not support it.

Change-Id: Ic42d25590ca8a836b29130138faba8d57cb03081