Fix pressing 'q' in the debugger console
This terminates the debugger, and thus closes the console, but we were
still trying to use it in Execute_Command. For some reason, the call
to Ref/In_Destruction/Unref does not seem to be enough anymore to
handle this.

The source of the behavior change is unclear (though it is related
to commit f4f35763aa730ecf170563b10a9c9ca18ff9634c). The workaround
put in place here is not to force the closing of MDI children when
the debugger terminates. This seems like the correct behavior in
any case, to preserve the position of the views that the user has
set.

There is however one visible change of behavior when the user starts
multiple debuggers, since the second debugger console and other
views are not closed when the second debugger is terminated, but the
first one is kept running. Again, this doesn't seem like a bad
behavior since that also better preserves the layout of windows
when using multiple debugger, something that the user might be
doing regularly.

P318-031 (1/3)

Change-Id: Ie228fb6186dbe85c7968ff2f93636216378982b1