Remove access to deallocated memory in Recent menu
Running one of the menus Project->Recent->x
causes a project reload, which in turn causes a recomputation
of this menu, thus freeing the command while it is running.
This can causes holders to this command to access unallocated
memory, and cause GPS to hang or crash while using the
menu.

Reported by customer under P310-027.

Change-Id: I31bb00c3b95e83fb525a51939d8f5149c704536e