NA09-049 (1/3) Preserve set of listeners for notification.
It is possible to register new listeners during notification. This patch
allows to notify only listners that was registered before start of
notification. All listeners added during notification are not called in
current notification cycle. It is possible to remove listeners during
notification too, this patch doesn't notify listeners that was removed
in current notification cycle.

Change-Id: I8807ec6df55d2ab6c38ea700af354e9ed409cf73