P407-018 fixing LaTeX doc generation
'\' is escaped in a python string, as a special caracter.
but it will be also escaped when compiled as "reSt".

To make sure that the actual "\b" is printed in the pdf,
"reSt" must receive the string "\\b" and to make sure "reSt" receive
the string "\\b", python string must be "\\\\b".

Change-Id: I99bb449290baa93faa242e6972168637654985f4