Enable the memory usage view only for 'default' build mode
P414-037

The '-largs, --print-memory-usage' switch is now automatically
disabled and removed from the command line if we are not in
the 'default' build mode.

This is done because non-default build modes usually append
extra arguments on the command line: as a result, some options
(e.g: the '--subdirs' option in the 'gnatcoverage' build mode)
can be interpreted as being part of the '-largs' section, leading
to a non-valid command line.

Change-Id: I49db44f7b4be3c8196a481df3f90c7d32e24ec5a