Add a new package to create dialog views
P427-007

A new package has been added in order to create dialog views.

This package defines types and associated subprograms that
allow to create dialog views that share a common look and feel.

This look and feel is the same as the one already used for the
preferences dialog. In particular, it includes:

   . The ability of gathering widgets within a view into groups

   . The ability to align each widget of the same view

   . The ability to add inlined documentation below each widget

   . The ability to display errors on widgets

Packages that implement the preferences dialog GUI have been adpated
to use this newly created package.

Change-Id: Ib1082c49b388fa80e3bd3efa443a57e8dee8d75b