Fix path to mdl2json
Rename modeling.py to qgen.py, to match the name of the product we are
interfacing with.

Use workflows instead of standard actions, to simplify the code, and enable
a "generate-code-then-build" build target. This makes working with code
generation more convenient.

Add a workflow for generate-build-debug

workflows: fix propagation of the last yield value from a sub-generator, as
in:
    def foo():
        yield 1
        yield 2
    def bar():
        st = yield foo()
        assert st == 2

workflows.make_action_from_workflow: removed. Instead, the usual
 gps_utils.make_interactive can be used since it is now compatible with
 generator functions as well.

We no longer generate the mapping file by parsing the sources, but
 instead rely on the mapping file generated by qgenc itself (requires a
 recent version of qgenc)


Part of P330-028 (1/2)

Change-Id: I69ded688e39c7571fec7ded39376221415a72306