This fixes searching for empty string.
In case of an empty string Context.Finish shouldn't be taken into
account, because it is -1 always. Otherwise Context.Finish >
Context.Start. So Context.Finish shouldn't be tested against
Context.Ref.

This issue was found in P511-024 in gpl-2016, but I'm able to reproduce
it on trunk manually (test driver works on truck however).

Candidate for merging in gpl-2016 and ide-2015

Change-Id: I9b3fd79ce3eb73cd6777d2d30e0a3f422e777d86