Add a runtime selector in the Project Properties
P401-030

P427-007 (1/3)

A runtime selector has been added in the Project Properties editor,
in the Toolchain page.

Scanning the available toolchains/runtimes is now automatically done
when launching the Project Properties editor. A spinner
is displayed while GPS is scanning, so that the user can see that GPS is
doing something in background.

Some initial changes have been made to the Project Properties editor UI
so that it used the Dialog_Package. More important changes will be made
in further commits.

Change-Id: I0e9f038386fba910ca389531ea6069b98c26f816