Relooking of the Project Properties editor
P427-007 (1/2)

The Project Properties editor has been relooked. It now has
the same look-and-feel as the Preferences editor dialog.

A new XML tag has been introduced for mutually exclusive
attributes, making it easier to specify them. The user is
now able to choose which mutually exclusive attribute he
wants to use with radio buttons, hiding the ones that
are not selected. 

Change-Id: Ib59cd5fe3750e36a5721770e3761fcdb1abdf203