Rewrite debugger interface to editors
We were still using code from the time we converted GVD, and that supported
multiple kinds of editors. Instead, we now provide a tighter integration
with the rest of GPS, which removes some extra levels of indirection.

Remove direct dependency on gtk+ for the editor interface in the debugger.

("debugger_location_changed"): new hook to report changes when the current
 location has changed (debugger stopped, user changed frame,...) This
 removes some direct dependencies between the debugger modules, and makes
 this hook usable for the QGen integration since we want to show the
 corresponding diagram then. In particular, there was no easy way, from
 python, to find when the user switches frame, and retrieving the
 "current file" information involved looking at the current editor buffer,
 which is a different kind of information.
Current file and line are set at the same time now, so that we can send
 signals only once.

QGen integration now properly updates the diagram when the user selects
 a new frame in the Call Stack window

continuation work for P330-028

Change-Id: I6de581f8a44bbe855c92c65f7e5b17960601b425