Tooltips now show path names relative to root project
In addition to showing absolute path names in notebook tabs, Windows
view and Project view, we now also display the path relative to the
root project, which can be much shorter in typical setups.

The code to compute the text of the tooltip is also shared between
all three cases.

Fix tooltip computation in Windows view, which was off by one line
since we introduced the local toolbar (a long time ago)

P428-038

Change-Id: Ifcb5b8a37b9a59701df7521c8d9c06370602928c