'Debug on Board': display errors if remote settings not set
P524-033

When clicking on the 'Debug on Board' button, verify that the
remote target and remote protocol settings have been set for
the spawned debugger.

If it's not the case, display an error message in the Messages
view indicating where the user can edit these settings in the
Project Properties editor.

Change-Id: I2de58e16456ebcd27adf8cfcbb5f21e4fc7dfe56