Add support for continuous signal overwriting in debugger
The support relies on the gdb python API using the
qgen_gdb.py script that is loaded when a debugger
is spawned.
This script defines 3 custom gdb commands to add and
delete watchpoints and to process them when the debugger
stops.

A watchdog breakpoint is created for each scope that has
watchpoints to allow the recreation of watchpoint after
the program execution goes out the said scope.

P526-017

Change-Id: I0f195dfd840dce3f3ff6e767f513d8a2b53813f1