Don't close Preferences dialog when pressing Escape on search
P415-012

The mechanism used to give the focus back to the previously focused
MDI child when pressing the ESCAPE key in the omnisearch has been
moved to the 'gtkada-entry_completion' package instead of
'GPS.Search.GUI'.

This allow us to share this mechanism for all the search bars, including
the one used in the Prefereces dialog.

Change-Id: I48f23d23df5087749f8852a50bee8e22f807fe9e