Proper handling of unfound executables in ProcessWrapper
P406-027

When GPS tries to spawn an external process with an unfound executable,
an error message is displayed in the Messages view and the 'on_exit'
callback is run if the process was spawned from Python.

In this case, we should not return any promise in
ProcessWrapper.wait_until_match but None instead (since the associated
process could not be spawned)

The st_link.py plug-in has been modified acoordingly.

Change-Id: I6afa280c2d21961cf15aaa1982b904c269ad4854