Improve qgen debugger navigation between subsystems
The previous block arrow to navigate to a parent has been replaced
by a toolbar button and a new previous button has been also added.
This is required to support library blocks and model references
that do not have a fixed parent but can be shared by multiple
blocks.
This patch does not require any change on the GPS side, a patch
updating mdl2json is also ready for the ptoolset repository.
It adds a 'qgen_navigation_info' item that is invisible for each
subsystem that contains a data field with a parent entry and a text
field updated to the previous subsystem id dynamically.
WIP : The buttons should be
grayed out when no parent or previous subsystem is available

P616-042

Change-Id: I8e0d32ac516419ca4b43befa3f1dc3cb68487595