Use a Gtk_Text_View for the 'Debug/Run..' window
P630-007

The Std_Dialogs package has been modified to provide the possibility
of having text input dialogs displaying a Gtk_Text_View instead of
a Gtk_Combo_Box_Text with an entry for text input.

Since GDB arguments can be quite long depending of the usage, we now use
this ability for the 'Debug/Run...' window.

Change-Id: Ib66380fb81e1423d2b35c0b3eae04d322c9e5501