Minor style adjustment.
The default style for the selected line seems to be overridden by
the colortheme.py (it looks at those values directly in the gtk+
theme). Nonetheless, our default style is a very light blue background
with white foreground, hard to read.
This patch makes the forground black like the rest of the text

Follow up on P630-019

Change-Id: I62cc6aae2fb7cb8e7a571f582db862e9348073fc