Allow showing the Breakpoints view even if no debugger is started
P712-016

Since we can now set breakpoints without a debugger, we need to be able
to see the view at all times.  In the /Debug/Data menu, all views are
now grayed out when no debugger is started (since the menu has no effect
anyway), except for the breakpoints, to make it clearer that the
Breakpoints view can be displayed.

This also fixes an issue we previously had: set a breakpoint outside of
the debugger, when /Debug/Initialize would empty the debugger view (and
marks on the sides of editors) until the debugger was actually started.

Remove the Code_Editor type, since every thing is in fact delegated to
the source editor anyway. This makes it possible to use the remaining
subprograms even when no debugger is started.

The "debugger_breakpoints_changed" hook can now be run even when no
debugger is running (so giving a null argument for its Debugger
parameter). This was already allowed, but most callbacks did not handle
this properly. This will eventually allow the QGen plugin to also
highlight blocks with breakpoints when the debugger is not running.

Change-Id: I8626157347d031de8dd5df872582abe2e25730c2