Bookmarks: move back to GPS's custom tooltips
The gtk+ tooltips require recomputing the contents of the tooltip (and,
internally, recreating the widgets) every time the mouse moves, which
is inefficient.
Our own implementation is better, although it required a minor improvement
to hide the tooltip when the user clicks (so that for instance the tooltip
does not appear during a drag operation in the Bookmarks view)

Tooltip: add a browse-mode, like gtk+ tooltips

This ensures that if a tooltip was displayed less than 500ms ago, we
display a new one much faster than otherwise. This makes it easier for
users to look at multiple tooltips.
We now use our own tooltips for the menus, since the gtk+ tooltips are
displayed such that it is too easy for the mouse to enter the tooltip,
and from that point on the menu items are no longer selectable until the
user closes the menu and opens it again.

P706-031

Change-Id: I6fb04c8b0e789ca48451024da5543eb32b25c528