Add a mode "Hide_Rather_Than_Close" to generic views
This allows keeping alive widgets even when they are
present in the MDI. Useful for keeping around the Vsearch
window, which we need for "find next" and "find previous"
actions.

Decorrelate Float_Child and Unfloat_Child, for clarity,
and adjust calls to Store_Position so that pressing ESC
to close the dialog will also store the position.

For P718-049.

Change-Id: I4fbe8fd578cff3793f96cf0b42adf4e936d8655d