P727-031 (1/2) GUI enhancements to the Omnisearch
Expose a Signal_Changed in the GtkAda.Entry_Completion. Use this
to store the history of the entry after each change: this allows
the user to press "backspace" to clear the default entry, and not
have the previous text be present next time the omnisearch is
brought up.

Use colors (the keywords foreground) to highlight matches
in the omnisearch results.

Make the Omnisearch view much larger.

Change-Id: Id45284a1d791e950d83e91c72236214a7917dc42