Add the notion of priority for enum preferences
P718-025

Add the notion of priority for enumeration preferences, like for the
other types of preferences.

When an enumeration preference has a limited number of choices, it is
displayed with radio buttons in a preferences group: in this case, we
also set the priority of the associated group with the preference's
priority since the group will only contain this preference.

Change-Id: I20667b67f37a87a90798ba1a4fad2bdbcb6c8859