P906-032: Refilter the Windows view every time it is refreshed
The Gtkada.Tree_View.On_Row_Inserted callback has been modified
so that it sets the visibility of the newly inserted row to True
by default: indeed, newly inserted rows are always null iters, so
it does not make sense to compute their visibility at this time.

As a result, a Gtkada.Tree_View needs to be refiltered every time
it is populated/refreshed when there is a filter pattern applied
on it.

Change-Id: Iaae9267db0bc82fa34028cc5e89b6062c39df8d5