Check the actual permissions on the disk when saving a file
If a file was read-only when it was opened, GPS would prevent its
saving otherwise, in particular when using "File/Save More/All"
(even if there is nothing modified in the file) and display an
error message.
G921-011 (1/2)

We now display a dialog asking whether the user wants to save anyway.
This is called hijacking in the context of Clearcase, and is necessary
for some users.
L404-006

Change-Id: I9b9f2239f10e1eacde28b79f7ea13a7c7ed5c28d