Omnisearch: preview should have same height
Change the height of the preview popup to match that of the omnisearch
popup, to display more information.

Add a small margin at the bottom. When in full screen, otherwise, users
might have the impression that the window extends below the screen, since
there are no borders.

Reduce margin between preview and popup

P920-006

Change-Id: Ic8b4c3e532d34b098fb7486460ec5d0a4552efe3